Update zh_cn.lang
I find that the names of blocks and items are still English when I change the language to Chinese in Minecaft. And I find the causes.
It's because you changed the name of file into zh_ch,lang . Now I have corrected some wrong translated names , you could add this new translation in the mod.