updated zh_cn.lang
I find the question of the names of blocks and items are still English when changing the language to Chinese in Minecraft. It's because that you changed the name of zh_cn.lang to zh_ch.lang. Now I have corrected some wrong translated names, you could add it to the mod.